LayoutTests/fast/regions/webkit-named-flow-event-add-to-flow.html

<html>
<head>
<script src="../../fast/js/resources/js-test-pre.js"></script>
<style>
#article {
    -webkit-flow-into: article;
    border: 2px solid black;
}
#region_1 {
    -webkit-flow-from: no_article;
    overflow:hidden;
    display:inline-block;
    vertical-align:top;
    margin:10px;
}
.block {
   display:block;
}
#region_1, #region_2 { height:110px; width:200px; margin: 5px;}
</style>
</head>
<body>
<div id="description" class="block"></div>
<div id="article">
    <div>
        <div id="content" style="height: 200px; background-color: green;"></div>
    </div>
</div>

<div id="region_1"></div>
<div id="region_2"></div>

<div id="console" class="block"></div>

<script>
    description("Tests regionLayoutEvent")
    if (window.layoutTestController) {
        layoutTestController.dumpAsText();
        layoutTestController.waitUntilDone();
    }
    window.jsTestIsAsync = true;
    function flowContent(region, flowName) {
        var r = document.getElementById(region);
        r.style.webkitFlowFrom = flowName;
    }
    
    var count = 0;
    function regionLayoutUpdated(event) {
        shouldBeEqualToString("event.target.name", "article");
        
        if (++count > 1) {
            event.target.removeEventListener("webkitRegionLayoutUpdate", regionLayoutUpdated);
            
            // this should be reached after region_2 is added to flow
            finishJSTest();
            return;
        }
        
        flowContent("region_2", "article"); // this will trigger a new round of events due to layout
    }

    function startTest() {
        var flow = document.webkitGetFlowByName("article");
        flow.addEventListener("webkitRegionLayoutUpdate", regionLayoutUpdated);
        
        debug("Flow content");
        flowContent("region_1", "article");
    }
    window.addEventListener("load", startTest);
</script>
<script src="../../fast/js/resources/js-test-post.js"></script>
</body>
</html>
